move `LSPLoop` code from `request_dispatch.cc` to `LSPLoop.cc`
